added span handler
