+++ src/instrumentation/common/utils.ts
@@ -1,8 +1,17 @@
+import * as path from 'path';
+import * as fs from 'fs'
 import { context as context_api, propagation, Context } from "@opentelemetry/api";
 import { RandomIdGenerator } from "@opentelemetry/sdk-trace-node";
-import { MONOCLE_SCOPE_NAME_PREFIX } from "./constants";
+import { MONOCLE_SCOPE_NAME_PREFIX, SCOPE_CONFIG_PATH, SCOPE_METHOD_FILE } from "./constants";
+import { consoleLog } from "../../common/logging";
 
 const scope_id_generator = new RandomIdGenerator();
+
+const http_scopes: Record<string, string> = {};
+
+export const get_http_scopes = () => {
+    return http_scopes;
+}
 
 function _generateScopeId(): string {
     return `${scope_id_generator.generateTraceId()}`;
@@ -15,7 +24,7 @@ export function setScopesInternal<A extends unknown[], F extends (...args: A) =>
     thisArg?: ThisParameterType<F>,
     ...args: A
 ) {
-    let updated_baggage_context  = updateBaggageContextWithScopes(baggage_context, scopes);
+    let updated_baggage_context = updateBaggageContextWithScopes(baggage_context, scopes);
     return context_api.with(updated_baggage_context, fn, thisArg, ...args);
 }
 
@@ -43,7 +52,7 @@ export function setScopesBindInternal(
     baggage_context: Context | null = null,
     fn: Function
 ): Function {
-    let updated_baggage_context  = updateBaggageContextWithScopes(baggage_context, scopes);
+    let updated_baggage_context = updateBaggageContextWithScopes(baggage_context, scopes);
     return context_api.bind(updated_baggage_context, fn);
 }
 
@@ -60,8 +69,7 @@ export function setScopesBindInternal(
 export function getScopesInternal(): Record<string, string> {
     const monocle_scopes: Record<string, string> = {};
     const currentBaggage = propagation.getActiveBaggage();
-    if (!currentBaggage)
-    {
+    if (!currentBaggage) {
         return monocle_scopes;
     }
 
@@ -83,6 +91,32 @@ export function getBaggageForScopes(): Context | null {
     }
     propagation.setBaggage(currentContext, baggage);
     return currentContext;
+}
+
+export function load_scopes(): any[] {
+    let methods_data: any[] = [];
+    let scope_methods: any[] = [];
+    if (!SCOPE_CONFIG_PATH) {
+        consoleLog('SCOPE_CONFIG_PATH not set');
+        return scope_methods;
+    }
+    try {
+        const methodsJson = fs.readFileSync(
+            path.join(SCOPE_CONFIG_PATH || '', SCOPE_METHOD_FILE),
+            'utf8'
+        );
+        methods_data = JSON.parse(methodsJson);
+        for (const method of methods_data) {
+            if (method.http_header) {
+                http_scopes[method.http_header] = method.scope_name;
+            } else {
+                scope_methods.push(method);
+            }
+        }
+    } catch (e) {
+        consoleLog(`Error loading scope methods from file: ${e}`);
+    }
+    return scope_methods;
 }
 
 // export function set_scopes_from_baggage(baggage_context: Context): void {
